% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_acronyms.R
\name{extract_acronyms}
\alias{extract_acronyms}
\title{Find acronyms in parentheses}
\usage{
extract_acronyms(txt)
}
\arguments{
\item{txt}{A tibble from pmc_text or character vector}
}
\value{
A tibble with acronyms
}
\description{
This function searches for words preceding the acronym that start with the
same initial letter and will likely fail in many situations.
}
\examples{
txt <- c(
  "An acronym like multinucleated giant cell (MGC)",
  "is later mentioned as MGC in the paper.")
extract_acronyms(txt)
}
\author{
Chris Stubben
}
