% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_string.R
\name{path_string}
\alias{path_string}
\title{Print a hierarchical path string}
\usage{
path_string(x, n)
}
\arguments{
\item{x}{a vector of names}

\item{n}{a vector of numbers with indentation level}
}
\value{
a character vector
}
\description{
Print a hierarchical path string from a vector of names and levels
}
\note{
Used by \code{\link{pmc_text}} to print full path to subsection title
}
\examples{
x <- c("carnivores", "bears", "polar", "grizzly", "cats", "tiger", "rodents")
n <- c(1,2,3,3,2,3,1)
path_string(x, n)

}
\author{
Chris Stubben
}
