% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_reference.R
\name{pmc_reference}
\alias{pmc_reference}
\title{Format references in PMC XML}
\usage{
pmc_reference(doc)
}
\arguments{
\item{doc}{\code{xml_document} from PMC}
}
\value{
a tibble with id, pmid, authors, year, title, journal, volume, pages,
and doi.
}
\description{
Format references in PMC XML
}
\note{
Mixed citations without any child tags are added to the author column.
}
\examples{
# doc <- europepmc::epmc_ftxt("PMC2231364")
doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
x <- pmc_reference(doc)
x

}
\author{
Chris Stubben
}
