% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_xml.R
\name{pmc_xml}
\alias{pmc_xml}
\title{Download PMC XML}
\source{
\url{https://github.com/ropensci/europepmc}
}
\usage{
pmc_xml(id)
}
\arguments{
\item{id}{a PMC id starting with 'PMC'}
}
\value{
\code{xml_document}
}
\description{
A wrapper for \code{epmc_ftxt} in the \code{europepmc} package to download
an \code{xml_document} from Europe PMC
}
\examples{
\dontrun{
doc <- europepmc::epmc_ftxt("PMC2231364") # OR
doc <- pmc_xml("PMC2231364")
}
doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
doc

}
