% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_refs.R
\name{separate_refs}
\alias{separate_refs}
\title{Separate references cited in full text into multiple rows}
\usage{
separate_refs(txt, pattern = "\\\\[[0-9, -]+\\\\]", column = "text")
}
\arguments{
\item{txt}{a table}

\item{pattern}{regular expression to match references, default "\\[[0-9, -]+\\]"}

\item{column}{column name, default "match"}
}
\value{
a tibble
}
\description{
Separates comma-delimited numeric strings and expands ranges like 7-9 into new rows
}
\examples{
x <- data.frame(row=1, text = "some important studies [7-9,15]")
separate_refs(x)

}
\author{
Chris Stubben
}
