% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypmc-package.R
\docType{package}
\name{tidypmc}
\alias{tidypmc-package}
\alias{tidypmc}
\title{\code{tidypmc} package}
\description{
Parse full text XML documents from PubMed Central
}
\details{
See the Github page for details at \url{https://github.com/ropensci/tidypmc}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/tidypmc}
  \item Report bugs at \url{https://github.com/ropensci/tidypmc/issues}
}

}
\author{
\strong{Maintainer}: Chris Stubben \email{chris.stubben@hci.utah.edu}

}
\keyword{internal}
