% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_caption.R
\name{pmc_caption}
\alias{pmc_caption}
\title{Split captions into sentences}
\usage{
pmc_caption(doc)
}
\arguments{
\item{doc}{\code{xml_document} from PMC}
}
\value{
a tibble with tag, label, sentence number and text
}
\description{
Split figure, table and supplementary material captions into sentences
}
\examples{
# doc <- epmc_ftp("PMC2231364") # OR
doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
x <- pmc_caption(doc)
x
dplyr::filter(x, sentence==1)

}
\author{
Chris Stubben
}
