% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_metadata.R
\name{pmc_metadata}
\alias{pmc_metadata}
\title{Get article metadata}
\usage{
pmc_metadata(doc)
}
\arguments{
\item{doc}{\code{xml_document} from PubMed Central}
}
\value{
a list
}
\description{
Get a list of journal and article metadata in /front tag
}
\examples{
# doc <- pmc_xml("PMC2231364") # OR
doc <- xml2::read_xml(system.file("extdata/PMC2231364.xml",
  package = "tidypmc"
))
pmc_metadata(doc)
}
\author{
Chris Stubben
}
