% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_table.R
\name{pmc_table}
\alias{pmc_table}
\title{Convert PMC XML tables into a list of tibble}
\usage{
pmc_table(doc)
}
\arguments{
\item{doc}{\code{xml_document} from PMC}
}
\value{
a list of tibbles
}
\description{
Convert PMC XML tables into a list of tibble
}
\note{
Saves the caption and footnotes as attributes and collapses multiline headers,
expands all rowspan and colspan attributes and adds
subheadings to column one.
}
\examples{
# doc <- epmc_ftxt("PMC2231364")
doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
x <- pmc_table(doc)
sapply(x, dim)
x
attributes(x[[1]])

}
\author{
Chris Stubben
}
