% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_text.R
\name{pmc_text}
\alias{pmc_text}
\title{Split section paragraphs into sentences}
\usage{
pmc_text(pmc)
}
\arguments{
\item{pmc}{\code{xml_document} from PMC}
}
\value{
a tibble with section, paragraph and sentence number and text
}
\description{
Split section paragraph tags into a table with subsection titles and sentences
using \code{tokenize_sentences}
}
\note{
Subsections may be nested to arbitrary depths and this function will return the
entire path to the subsection title as a delimited string, eg. "Results; Predicted
functions; Pathogenicity".  Any tables, figures and formulas that are nested in
section paragraphs are removed.
}
\examples{
# doc <- europepmc::epmc_ftxt("PMC2231364")
doc <- read_xml(system.file("extdata/PMC2231364.xml", package = "tidypmc"))
txt <- pmc_text(doc)
txt
dplyr::count(txt, section, sort=TRUE)

}
\author{
Chris Stubben
}
