% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc_xml.R
\name{pmc_xml}
\alias{pmc_xml}
\title{Download XML from PubMed Central}
\source{
\url{https://europepmc.org/RestfulWebService}
}
\usage{
pmc_xml(id)
}
\arguments{
\item{id}{a PMC id starting with 'PMC'}
}
\value{
\code{xml_document}
}
\description{
Download XML from PubMed Central
}
\examples{
\dontrun{
doc <- pmc_xml("PMC2231364")
}

}
