% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_refs.R
\name{separate_refs}
\alias{separate_refs}
\title{Separate references cited into multiple rows}
\usage{
separate_refs(txt, column = "text")
}
\arguments{
\item{txt}{a table}

\item{column}{column name, default "text"}
}
\value{
a tibble
}
\description{
Separates references cited in brackets or parentheses into multiple rows and splits
the comma-delimited numeric strings and expands ranges like 7-9 into new rows
}
\examples{
x <- data.frame(row=1, text = "some important studies [7-9,15]")
separate_refs(x)

}
\author{
Chris Stubben
}
