% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypmc-package.R
\docType{package}
\name{tidypmc}
\alias{tidypmc}
\alias{tidypmc-package}
\title{\code{tidypmc} package}
\description{
Parse full text XML documents from PubMed Central
}
\details{
See the Github page for details at \url{https://github.com/ropensci/tidypmc}
}
\keyword{internal}
